+++ packages/docx/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @udecode/plate-docx-serializer
 
+## 38.0.6
+
+## 38.0.5
+
 ## 38.0.1
 
 ### Patch Changes